migrate Itinerary to styled-components
